% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLineup.R, R/formation.R
\name{getLineup}
\alias{getLineup}
\alias{getLineup,formation-method}
\title{Lineup of a united formation}
\usage{
getLineup(obj)

\S4method{getLineup}{formation}(obj)
}
\arguments{
\item{obj}{object of the class \code{formation}.}
}
\value{
\code{vector} of the used lineup
}
\description{
Generates a \code{numeric} vector which specifies the used united lineup
}
