% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overtime.R
\name{overtime}
\alias{overtime}
\title{Computing overtime results}
\usage{
overtime(chancesHome, chancesAway, probGoalHome, probGoalAway)
}
\arguments{
\item{chancesHome}{goalscoring chances of home team}

\item{chancesAway}{goalscoring chances of away team}

\item{probGoalHome}{probability of scoring a goal for home team}

\item{probGoalAway}{probability of scoring a goal for away team}
}
\value{
\code{list} with probabilities of final outcome.
}
\description{
Computes the final overtime outcome.
}
