% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penaltyGoalsProb.R
\name{penaltyGoalsProb}
\alias{penaltyGoalsProb}
\title{Computing goals by united}
\usage{
penaltyGoalsProb(posPenalties, penaltyGoalProb, penaltyProb = 0.1)
}
\arguments{
\item{posPenalties}{number of possible penalties in a game}

\item{penaltyGoalProb}{probability of a goal by a singular penalty}

\item{penaltyProb}{occurrence probability of a penalty}
}
\value{
A \code{data.frame} with two columns: the possible goals and the probability
for achieving this number of goals.
}
\description{
Computes the distribution function of possible goals by penalties.
}
