% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overtime.R
\name{penaltyShootout}
\alias{penaltyShootout}
\title{Computing outcome of penalty shootout}
\usage{
penaltyShootout(probPenaltySaveHome, probPenaltySaveAway, initial = 5)
}
\arguments{
\item{probPenaltySaveHome}{probability of saving a penalty for home team}

\item{probPenaltySaveAway}{probability of saving a penalty for away team}

\item{initial}{number of initial penalties (default 5)}
}
\value{
\code{list} with probabilities of final outcome (\code{winProbabilityHome}, 
\code{winProbabilityAway}).
}
\description{
Computes outcome of a penalty shootout.
}
