% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simRedCard.R, R/formation.R
\name{simRedCard}
\alias{simRedCard}
\alias{simRedCard,formation,numeric,matrix-method}
\title{Simulate red card(s)}
\usage{
simRedCard(obj, lineup, Hard)

\S4method{simRedCard}{formation,numeric,matrix}(obj, lineup, Hard)
}
\arguments{
\item{obj}{object of the class \code{formation}}

\item{lineup}{lineup of the corresponding object \code{obj}}

\item{Hard}{matrix of hardness to be used}
}
\value{
\code{list} with two elements: 
\itemize{
 \item{\code{vector}}{ adjusted lineup for the red card(s)}
 \item{\code{numeric}}{ number of red cards}
}
}
\description{
Simulates red card(s) in the united and returns the adjusted lineup.
}
