% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitedRPackage.R
\docType{package}
\name{unitedR-package}
\alias{unitedR-package}
\alias{unitedR}
\title{Assessment and Evaluation of United Formations}
\description{
Assessment and Evaluation of United Formations
}
\details{
\tabular{ll}{
Package: \tab unitedR \cr
Type: \tab Package\cr
Version: \tab 0.3.1 \cr
Date: \tab 2018-01-03 \cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes \cr
}

This package provides functionality for the assessment of lineups and formations in United. The rules for
United in detail can be found under: \href{http://www.schroepl.net/pbm/software/united/_dokument/handbuch.htm}{United-rules}.
}
\references{
\href{http://www.omido.de/}{omido}, \href{http://www.schroepl.net/pbm/software/united/}{United Software}, \href{http://schroepl.net/projekte/united_forum/}{United-Forum}
}
\author{
David Schindler \email{dv.schindler@gmail.com}
}
